update card add

+++ game-api/src/main/java/com/example/gameapi/service/RoleService.java
@@ -5,5 +5,5 @@ import com.example.gameapi.dto.RoleDto;
 import java.util.List;
 
 public interface RoleService {
-  List<RoleDto> getRoles(int gameId);
+  List<RoleDto> getRoles(Long gameId);
 }
